refactor: ♻️ rename selectTags2 and selectLocation2
