+++ pages/project/[id]/edit/location.tsx
@@ -18,7 +18,7 @@ import EditProjectLayout from "components/layout/EditProjectLayout";
 import FetchProjectLayout, { useProject } from "components/layout/FetchProjectLayout";
 import Layout from "components/layout/Layout";
 import EditFormLayout from "components/partials/project/edit/EditFormLayout";
-import { SelectedLocation } from "components/SelectLocation2";
+import { SelectedLocation } from "components/SelectLocation";
 import { GetStaticPaths } from "next";
 
 //
